+++ testV4.py
@@ -2,8 +2,10 @@ import shutil
 import os
 import zipfile
 import clean
-shutil.rmtree("downloads")
-shutil.rmtree("sorted")
+if os.path.exists("downloads"):
+    shutil.rmtree("downloads")
+if os.path.exists("sorted"):
+    shutil.rmtree("sorted")
 os.makedirs("sorted")
 os.makedirs("sorted/TV_shows")
 os.makedirs("sorted/unrecognized")
